script: Move files into `script/{css,fetch}`
Testing: it compiles
