+++ components/script/dom/document/document.rs
@@ -81,6 +81,8 @@ use time::Duration as TimeDuration;
 use url::{Host, Position};
 
 use crate::animations::Animations;
+use crate::css::stylesheet_loader::StylesheetContextId;
+use crate::css::stylesheet_set::StylesheetSetRef;
 use crate::dom::FlatTreeParent;
 use crate::dom::animationtimeline::AnimationTimeline;
 use crate::dom::attr::Attr;
@@ -209,12 +211,10 @@ use crate::dom::xpathevaluator::XPathEvaluator;
 use crate::dom::xpathexpression::XPathExpression;
 use crate::event_loop::document_loader::{DocumentLoader, LoadType};
 use crate::event_loop::script_thread::{ScriptThread, SharedRwLocks};
-use crate::fetch::{DeferredFetchRecordInvokeState, FetchCanceller};
+use crate::fetch::fetch::{DeferredFetchRecordInvokeState, FetchCanceller};
+use crate::fetch::network_listener::{FetchResponseListener, NetworkListener};
 use crate::mime::{APPLICATION, CHARSET};
 use crate::navigation::navigate;
-use crate::network_listener::{FetchResponseListener, NetworkListener};
-use crate::stylesheet_loader::StylesheetContextId;
-use crate::stylesheet_set::StylesheetSetRef;
 use crate::tasks::task::NonSendTaskBox;
 use crate::tasks::task_manager::TaskManager;
 use crate::tasks::task_source::TaskSourceName;
