+++ components/script/dom/eventsource.rs
@@ -40,8 +40,10 @@ use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::messageevent::MessageEvent;
 use crate::dom::performance::performanceresourcetiming::InitiatorType;
-use crate::fetch::{FetchCanceller, RequestWithGlobalScope, create_a_potential_cors_request};
-use crate::network_listener::{self, FetchResponseListener, ResourceTimingListener};
+use crate::fetch::fetch::{
+    FetchCanceller, RequestWithGlobalScope, create_a_potential_cors_request,
+};
+use crate::fetch::network_listener::{self, FetchResponseListener, ResourceTimingListener};
 use crate::realms::enter_auto_realm;
 use crate::timers::OneshotTimerCallback;
