+++ components/script/dom/html/embedded_content/htmlimageelement.rs
@@ -67,9 +67,9 @@ use crate::dom::srcset::SourceSet;
 use crate::dom::window::Window;
 use crate::event_loop::document_loader::{LoadBlocker, LoadType};
 use crate::event_loop::script_thread::ScriptThread;
-use crate::fetch::{RequestWithGlobalScope, create_a_potential_cors_request};
+use crate::fetch::fetch::{RequestWithGlobalScope, create_a_potential_cors_request};
+use crate::fetch::network_listener::{self, FetchResponseListener, ResourceTimingListener};
 use crate::microtask::MicrotaskRunnable;
-use crate::network_listener::{self, FetchResponseListener, ResourceTimingListener};
 use crate::realms::enter_auto_realm;
 
 /// <https://html.spec.whatwg.org/multipage/#img-req-state>
