+++ components/script/dom/html/embedded_content/htmlvideoelement.rs
@@ -44,8 +44,8 @@ use crate::dom::node::virtualmethods::VirtualMethods;
 use crate::dom::node::{Node, NodeDamage, NodeTraits};
 use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::event_loop::document_loader::{LoadBlocker, LoadType};
-use crate::fetch::{FetchCanceller, RequestWithGlobalScope};
-use crate::network_listener::{self, FetchResponseListener, ResourceTimingListener};
+use crate::fetch::fetch::{FetchCanceller, RequestWithGlobalScope};
+use crate::fetch::network_listener::{self, FetchResponseListener, ResourceTimingListener};
 use crate::url::ensure_blob_referenced_by_url_is_kept_alive;
 
 #[dom_struct]
