+++ components/script/dom/security/cspviolationreporttask.rs
@@ -26,8 +26,10 @@ use crate::dom::performance::performanceresourcetiming::InitiatorType;
 use crate::dom::reporting::reportingobserver::ReportingObserver;
 use crate::dom::securitypolicyviolationevent::SecurityPolicyViolationEvent;
 use crate::dom::types::GlobalScope;
-use crate::fetch::{RequestWithGlobalScope, create_a_potential_cors_request};
-use crate::network_listener::{FetchResponseListener, ResourceTimingListener, submit_timing};
+use crate::fetch::fetch::{RequestWithGlobalScope, create_a_potential_cors_request};
+use crate::fetch::network_listener::{
+    FetchResponseListener, ResourceTimingListener, submit_timing,
+};
 use crate::tasks::task::TaskOnce;
 
 pub(crate) struct CSPViolationReportTask {
