+++ components/script/dom/servoparser/mod.rs
@@ -92,8 +92,8 @@ use crate::dom::text::Text;
 use crate::dom::types::{HTMLElement, HTMLMediaElement, HTMLOptionElement};
 use crate::event_loop::document_loader::{DocumentLoader, LoadType};
 use crate::event_loop::script_thread::ScriptThread;
+use crate::fetch::network_listener::FetchResponseListener;
 use crate::navigation::determine_the_origin;
-use crate::network_listener::FetchResponseListener;
 use crate::realms::enter_auto_realm;
 use crate::script_runtime::IntroductionType;
