+++ components/script/dom/servoparser/prefetch.rs
@@ -27,7 +27,7 @@ use crate::dom::bindings::trace::{CustomTraceable, JSTraceable};
 use crate::dom::document::Document;
 use crate::dom::html::htmlscriptelement::script_fetch_request;
 use crate::dom::processingoptions::determine_cors_settings_for_token;
-use crate::fetch::create_a_potential_cors_request;
+use crate::fetch::fetch::create_a_potential_cors_request;
 use crate::modules::script_module::ScriptFetchOptions;
 
 #[derive(JSTraceable, MallocSizeOf)]
