+++ components/script/fetch/network_listener.rs
